Destructor deployMicroservice
